yolo_server/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
numpy
python-multipart
